pyproject.toml
[project]
name = "gearbox-showcase"
version = "0.1.0"
description = "Gearbox Protocol Analysis & Simulator"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "web3>=6.15.0",
    "eth-account>=0.10.0",
    "python-dotenv>=1.0.0",
]
